Downloads/YouTube Workspace (DOE)/Telegram/Bot/satmakon-bot/requirements.txt
python-telegram-bot==21.6
openai>=2.8.0
python-dotenv==1.0.1
gspread==6.1.2
google-auth>=2.47.0
